requirements.txt
discord.py>=2.3.0
ollama>=0.4.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
